backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
websockets
pytest